Avoid cross-configuration
Using `subprojects { ... }` block to cross-configure Gradle builds is
considered a bad practice. The rationale is that it makes it difficult
to understand what project applies what plugins and makes it even more
difficult to extract subprojects into their own full blown projects.

Last but not least, this is also bad for performance. As a consequence,
this commit updates the template to avoid that pattern, and use the
new `plugins` block. Note that there are other changes we can make to
make things easier. In particular, the whole `dependencies` block of
the main project should be extracted into some kind of internal Micronaut
module plugin.