Fix micronaut-core docs publishing (#202)
This PR allows the release workflow to work for both micronaut-core and all the other repos.

In the case of core, the docs must be published in the GitHub Pages of micronaut-docs. This
is defined in the
[`docsRepository`](https://github.com/micronaut-projects/micronaut-core/blob/3.4.x/gradle.properties#L6)
variable in core. For the other repos, they just publish docs to their own GitHub Pages,
which is represented by the `github.repository` context property.

Fixes https://github.com/micronaut-projects/micronaut-core/issues/7126