Enable javadoc parsing for UnusedImports checkstyle rule (#315)
Right now, when you reference a class in javadoc, you have to use the fully qualified name unless the class has already been imported. If you add an import only for the javadoc, checkstyle would complain that the import is unused.

After this change, checkstyle finds uses of the import in the javadoc properly.