fix: generate SLSA provenances for main artifacts
Some projects produce a large number of artifacts and
producing provenances for both artifacts and metadata files
results in a base64-subjects text that can easily grow larger
than the allowed size of command-line argument and environment strings.

To avoid hitting allowed size issues, as a workaround we generate
provenances for the main artifacts and POM files only.

Signed-off-by: behnazh-w <behnaz.hassanshahi@oracle.com>