Install both Java 21 and Java 17 in release workflows
This commit changes the release workflows so that the use Java 21 for publishing.
This will _not_ change the default version of Micronaut project to 17, this is
only installing Java 21, so that we can release modules which depend on Java 21,
such as `micronaut-langchain4j-jllama`.

In such a case, currently it's possible to configure a module to build with 21,
but it would never be released, because our release workflows depend on 17.
With this change, the module will now be released with 21. For modules which
are using the defaults, the Java 17 toolchain is going to be used.