Use Maven Central Publisher API
This commit updates the template to use the new publishing task, which
makes use of the Publisher API. This API is more reliable than the old
one since Maven Central changed their backend.

However, this requires that projects update their root project's build
script to use the new "parent" plugin, as seen in this PR.

In other words, publishing will fail if that plugin is not applied.

In addition, this reverts Gradle to 8.14.x, since our builds are NOT
ready for Gradle 9 (in particular any project which uses a Kotlin 1.9
plugin will FAIL).