Remove default JDKs (#608)
This commit removes the default JDKs which are installed when using
the ubuntu image. This causes issues with Gradle builds, where in
a workflow that uses Java 21, the test resources service may start
using Java 17 (but there are probably more occurrences). The safest
is to delete all installations prior to the setup java/graalvm
steps.

This may help surface more build configuration problems, such as
builds which are supposed to test Java 21 but are actually running
with 17. This is typically the case with some projects which use
the "public" Micronaut Gradle plugin in a test project, where the
default is to use a Java 17 toolchain.

Before this change, the test projects would accidentally use Java 17
instead of failing with a proper error message saying that no Java 17
toolchain is found (which is expected as we would want it to test 21).